Refactor logging utils and fix tests
