+++ requirements.txt
@@ -8,3 +8,4 @@ pandas_ta==0.3.14b0
 openpyxl>=3.1
 plotly>=5
 kaleido>=0.2
+hypothesis~=6.99
