+++ tests/test_indicator_calculator.py
@@ -6,6 +6,10 @@ import numpy as np
 import pandas as pd
 import pytest
 
+ta = pytest.importorskip("pandas_ta")
+if not hasattr(ta, "psar"):
+    pytest.skip("psar not available")
+
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 sys.modules.pop("pandas_ta", None)  # Ensure real pandas_ta is used in this test
